refactor: перенос логики добавления заказа в ingredientSlice.js
